requirement.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
fonttools==4.50.0
gunicorn==21.2.0
importlib_resources==6.3.1
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.5
matplotlib==3.8.3
mysqlclient==2.2.4
networkx==2.8
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.21
PyMySQL==1.1.0
pyope==0.2.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
radar==0.3
redis==5.0.3
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.18.1